Break these apart now that I understand how
